project/requirements.txt
fastapi==0.61.0
uvicorn==0.11.5
asyncpg==0.20.1
tortoise-orm==0.16.14
pytest==5.4.2
requests==2.23.0
gunicorn==20.0.4
pytest-cov==2.9.0
flake8===3.8.2
black==19.10b0
isort==4.3.21
passlib==1.7.2
bcrypt==3.1.7
python-jose==3.1.0
python-multipart==0.0.5